feat: Change server to accept connections forever
- Do not close server on timeout anymore, timeouts will only be for clients. This will avoid clients from constantly restarting the server.
Also, if the server is ever hosted in the cloud, clients would not have access to it to restart it.
Hence, it is kept alive forever unless it is closed by a user with access.

- Start start_game_when_two_clients thread whenever remove_client is called except when client's reply for play_again is 'no',
and other client has not yet entered play_again reply.
This is because if the other player has not entered their play_again reply, and the start_game_when_two_clients
thread starts, the main_game_thread waiting for the reply never receives one but receives a "status" msg
next because the start_game_when_two_clients thread has started again.
Another case where the thread will not start is when it is already running.

- Print other_client_disconnected msg always as long as the client or opponent's play_again reply is 'yes'