fix: clear errors only when game started and round is over
errors will only be cleared when game has started and the round is over
in order to display the final result. Code to validate ip address is
removed as it is now unnecessary since automatic server discovery is
now used to connect to server. Show non-socket error animation when
general Exception is caught.